mirgate functional and utils to utils.operations
