+++ torecsys/__init__.py
@@ -5,14 +5,10 @@ __version__ = "0.0.1-dev"
 
 import torecsys.data
 import torecsys.estimators
-import torecsys.functional
 import torecsys.inputs
 import torecsys.layers
 import torecsys.losses
+import torecsys.metrics
 import torecsys.models
+import torecsys.trainer
 import torecsys.utils
-
-from .metrics import metrics
-from .utils.training.ranking_trainer import RankingTrainer
-from .utils.training.trainer import Trainer
-
