+++ torecsys/data/sampledata/load_data.py
@@ -2,7 +2,6 @@ from . import __ml_size__
 import pandas as pd
 import os
 
-
 def load_ml_data(size : str,
                  dir  : str = None) -> pd.DataFrame:
     r"""Load movielens dataset from ml-[size]/ratings.csv file in ../sample_data to pandas.DataFrame.
