--- torecsys/data/sampledata/sample_data/.gitignore
@@ -1,2 +1,0 @@
-# exclude everything
-*/
