+++ torecsys/estimators/__init__.py
@@ -1,14 +1,14 @@
 r"""[In development] torecsys.estimators is a sub module of estimators, which can be called driectly with a fitted dataloader of inputs
 """
 
-from torecsys.utils.logging import TqdmHandler
-from torecsys.utils.training.trainer import Trainer
 from logging import Logger
 from os import path
 from pathlib import Path
 from texttable import Texttable
 import torch
 import torch.utils.data
+from torecsys.trainer import Trainer
+from torecsys.utils.logging import TqdmHandler
 from typing import Dict
 import warnings
