--- torecsys/functional/__init__.py
@@ -1,8 +1,0 @@
-r"""torecsys.functional is a sub module of functions' implementation which is shared by different kinds of models.
-"""
-
-from .attention import show_attention
-from .attention import dummy_attention
-from .inner_product_similarity import inner_product_similarity
-from .regularization import regularize
-from .regularization import Regularizer
