+++ torecsys/metrics/__init__.py
@@ -1,2 +1,3 @@
-r"""torecsys.metrics is a sub module to implemente functions to calculate metrcis of evaluateion in recommendation system
+r"""torecsys.metrics is a sub module to implemente functions to calculate metrcis of 
+evaluateion in recommendation system.
 """
