+++ torecsys/models/ctr/fat_deep_ffm.py
@@ -1,11 +1,10 @@
 from . import _CtrModel
-from torecsys.layers import CENLayer, FFMLayer, DNNLayer
-from torecsys.utils.decorator import jit_experimental, no_jit_experimental_by_namedtensor
-from torecsys.utils.utils import combination
 import torch
 import torch.nn as nn
+from torecsys.layers import CENLayer, FFMLayer, DNNLayer
+from torecsys.utils.decorator import no_jit_experimental_by_namedtensor
+from torecsys.utils.operations import combination
 from typing import Callable, List
-
 
 class FieldAttentiveDeepFieldAwareFactorizationMachineModel(_CtrModel):
     r"""Model class of Field Attentive Deep Field Aware Factorization Machine (Fat DeepFFM).
