+++ torecsys/models/ctr/product_neural_network.py
@@ -1,9 +1,9 @@
 from . import _CtrModel
-from torecsys.layers import InnerProductNetworkLayer, OuterProductNetworkLayer, DNNLayer
-from torecsys.utils.decorator import jit_experimental, no_jit_experimental_by_namedtensor
-from torecsys.utils.utils import combination
 import torch
 import torch.nn as nn
+from torecsys.layers import InnerProductNetworkLayer, OuterProductNetworkLayer, DNNLayer
+from torecsys.utils.decorator import no_jit_experimental_by_namedtensor
+from torecsys.utils.operations import combination
 from typing import Callable, List
 
 class ProductNeuralNetworkModel(_CtrModel):
