+++ torecsys/models/sequential.py
@@ -1,9 +1,8 @@
-from torecsys.inputs.inputs_wrapper import InputsWrapper
-from torecsys.models import _Model
 import torch
 import torch.nn as nn
+from torecsys.inputs.inputs_wrapper import InputsWrapper
+from torecsys.models import _Model
 from typing import Dict
-
 
 class Sequential(nn.Module):
     r"""Sequential container, where the module of embeddings and model will be added to it 
@@ -12,17 +11,17 @@ class Sequential(nn.Module):
     def __init__(self, 
                  inputs_wrapper : InputsWrapper, 
                  model          : _Model):
-        r"""Initialize Sequential,
+        r"""Initialize Sequential container.
         
         Args:
-            inputs_wrapper (torecsys.inputs.inputs_wrapper.InputsWrapper): Inputs wrapper where 
-                the return is a dictionary of inputs' tensors which are passed to the model directly.
-            model (torecsys.models._Model): Model class to be trained and used in prediction.
+            inputs_wrapper (InputsWrapper): Inputs wrapper where the return is a dictionary 
+                of inputs' tensors which are passed to the model directly.
+            model (_Model): Model class to be trained and used in prediction.
 
         Attributes:
-            inputs_wrapper (torecsys.inputs.inputs_wrapper.InputsWrapper): Inputs wrapper where 
-                the return is a dictionary of inputs' tensors which are passed to the model directly.
-            model (torecsys.models._Model): Model class to be trained and used in prediction.
+            inputs_wrapper (InputsWrapper): Inputs wrapper where the return is a dictionary 
+                of inputs' tensors which are passed to the model directly.
+            model (_Model): Model class to be trained and used in prediction.
         """
         # refer to parent class
         super(Sequential, self).__init__()
@@ -35,8 +34,8 @@ class Sequential(nn.Module):
         r"""Forward calculation of Sequential.
         
         Args:
-            inputs (Dict[str, T]): Dictionary of inputs, where key is name of input fields, and value 
-                is tensor pass to Input class.
+            inputs (Dict[str, T]): Dictionary of inputs, where key is name of input fields, 
+                and value is tensor pass to Input class.
         
         Returns:
             T: Output of model.
@@ -46,5 +45,6 @@ class Sequential(nn.Module):
 
         # calculate forward propagation of model
         outputs = self.model(**inputs)
+        
         return outputs
