--- torecsys/utils/decorator/__init__.py
@@ -1,4 +1,0 @@
-r"""torecsys.utils.decorator is a sub module of tools for decorator
-"""
-
-from .decorator import *
