--- torecsys/utils/logging/__init__.py
@@ -1,4 +1,0 @@
-r"""torecsys.utils.logging is a sub module of tools for logging
-"""
-
-from .tqdm import *
