--- torecsys/utils/training/__init__.py
@@ -1,5 +1,0 @@
-r"""torecsys.utils.training is a sub module of training tools, including Traniner
-"""
-
-from .ranking_trainer import RankingTrainer
-from .trainer import Trainer
