--- torecsys/utils/utils/__init__.py
@@ -1,6 +1,0 @@
-r"""torecsys.utils.utils is a sub module of functions in utils which cannot be grouped
-"""
-
-from .combination import combination
-
-ncr = combination
